reorganize the data folder
