+++ src/data/settings/index.ts
@@ -1,0 +1,4 @@
+export * from "./headerClickBehaviorSetting";
+export * from "./animatedExpandSetting";
+export * from "./repoHeaderBadgeSetting";
+export * from "./statusCheckSetting";
